# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchReviewRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "review_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Review Ids",
            "type": "array"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "review_ids",
          "reviewer"
        ],
        "title": "BatchReviewRequest",
        "type": "object"
      },
      "BatchReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "negative_rule_count": {
            "default": 0,
            "title": "Negative Rule Count",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          },
          "review_ids": {
            "items": {
              "type": "string"
            },
            "title": "Review Ids",
            "type": "array"
          }
        },
        "required": [
          "processed",
          "review_ids"
        ],
        "title": "BatchReviewResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "create_knowledge_candidate": {
            "default": false,
            "title": "Create Knowledge Candidate",
            "type": "boolean"
          },
          "reviewer": {
            "default": "demo_user",
            "title": "Reviewer",
            "type": "string"
          }
        },
        "title": "ReviewDecisionRequest",
        "type": "object"
      },
      "ReviewGroupItem": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "review_ids": {
            "items": {
              "type": "string"
            },
            "title": "Review Ids",
            "type": "array"
          }
        },
        "required": [
          "key",
          "count",
          "review_ids"
        ],
        "title": "ReviewGroupItem",
        "type": "object"
      },
      "ReviewGroupedResponse": {
        "additionalProperties": false,
        "properties": {
          "group_by": {
            "title": "Group By",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewGroupItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "group_by",
          "items"
        ],
        "title": "ReviewGroupedResponse",
        "type": "object"
      },
      "ReviewImpactItem": {
        "additionalProperties": false,
        "properties": {
          "confidence_after": {
            "title": "Confidence After",
            "type": "number"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "target_field": {
            "title": "Target Field",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "doc_id",
          "source_label",
          "target_field",
          "confidence_after"
        ],
        "title": "ReviewImpactItem",
        "type": "object"
      },
      "ReviewImpactPreview": {
        "additionalProperties": false,
        "properties": {
          "badcase_hits": {
            "items": {
              "type": "string"
            },
            "title": "Badcase Hits",
            "type": "array"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "risk_flags": {
            "items": {
              "type": "string"
            },
            "title": "Risk Flags",
            "type": "array"
          },
          "would_affect": {
            "items": {
              "$ref": "#/components/schemas/ReviewImpactItem"
            },
            "title": "Would Affect",
            "type": "array"
          }
        },
        "required": [
          "review_id"
        ],
        "title": "ReviewImpactPreview",
        "type": "object"
      },
      "ReviewListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewRecordResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
      "ReviewRecordResponse": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "doc_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "review_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Comment"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "schema_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Id"
          },
          "source_field_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Field Name"
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Path"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "suggested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested By"
          },
          "target_field_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Field Id"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "review_id",
          "task_id",
          "status",
          "reviewer",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewRecordResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SchemaPack Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "list_reviews_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/batch-approve": {
      "post": {
        "operationId": "batch_approve_reviews_api_v1_reviews_batch_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Approve Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/batch-reject": {
      "post": {
        "operationId": "batch_reject_reviews_api_v1_reviews_batch_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Reject Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/grouped": {
      "get": {
        "operationId": "grouped_reviews_api_v1_reviews_grouped_get",
        "parameters": [
          {
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "schema_id",
              "title": "Group By",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewGroupedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grouped Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/api/v1/reviews/{review_id}": {
      "get": {
        "operationId": "get_review_api_v1_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{review_id}/approve": {
      "post": {
        "operationId": "approve_review_api_v1_reviews__review_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{review_id}/impact-preview": {
      "post": {
        "operationId": "impact_preview_api_v1_reviews__review_id__impact_preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewImpactPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Impact Preview",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{review_id}/reject": {
      "post": {
        "operationId": "reject_review_api_v1_reviews__review_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRecordResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Review",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}